Get IDirect3dDevice from LearningModelDevice Closes #175
